+++ src/lib.rs
@@ -1,6 +1,7 @@
 #[doc(inline)]
-pub use self::{clients::Clients, mgmt::ManagementApi, users::Users};
+pub use self::{auth::AuthenticationApi, clients::Clients, mgmt::ManagementApi, users::Users};
 
+pub mod auth;
 pub mod clients;
 mod mgmt;
 pub mod models;