+++ CHANGELOG.md
@@ -8,3 +8,4 @@
   client detail.
 - **[Management API - Users]** Implemented endpoints list users and get user
   detail.
+- **[Authentication API]** Implemented client credentials flow.